requirements.txt

# SolarVer – Dependencias del backend
# Instalar con: pip install -r requirements.txt

 
flask==3.0.0
flask-cors==4.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
bcrypt==4.1.2
pytz==2024.1
APScheduler==3.10.4
pandas==2.0.0
openpyxl==3.1.0
reportlab==4.1.0
phonenumbers==8.13.0
dnspython==2.6.0
requests==2.31.0